Fix wrong rule
According to [Wikipedia](https://en.wikipedia.org/wiki/Reversi#Rules),
if one player can not make a valid move, play passes back to the other
player. When neither player can move, the game ends.
But the original program did a wrong policy that if player can not make
a valid move, then it raises `NoMovesError` and immediately ends the
game.